Unify Screw Breech Directionality
adds two direction conversions so screw breeches always have the same behaviour, that being unscrewing when fed counter-clockwise rotation and screwing in when fed clockwise rotation regardless of placement direction

current behaviour is that clockwise rotation will screw in when facing north or west, while unscrewing when facing south or east.

the second convertToDirection for partialTicks is to prevent a visual bug where the screw will jitter frantically while facing east or south.

there's a visual bug induced by this change in behaviour where feeding rotation when the screw breech is facing east or south where the screw visually will rotate opposite of the input rotation, as well as when the screw breech is vertical facing upwards.

I'll also note here: 
this is quite literally the first time I've proposed anything to a github project, at most before this I've screwed around with JSON editing and literally only learned how to use git bash and gradle today. there's probably a more efficient way to do this but I kept getting errors tossed at me and this was the first result that worked fully without a flagrant visual bug.